second

+++ Shauli_blog/Controllers/MyController.cs
@@ -3,69 +3,50 @@ using System;
 using System.Collections.Generic;
 using System.Data;
 using System.IO;
-using System.Linq;
-using System.Net;
 using System.Web;
 using System.Web.Mvc;
+using System.Web.Security;
+using DotNetOpenAuth.AspNet;
+using System.Linq;
+using System.Net;
 using WebMatrix.WebData;
 
 namespace Shauli_blog.Controllers
 {
+    
     public class MyController : Controller
     {
         UsersContext db = new UsersContext();
-
-        // GET: /My/
-        [HttpGet]
-        public ActionResult Index()
-        {
-            
-            ViewBag.UserName = User.Identity.Name;
-            var CArticle =db.Articles.ToList().OrderByDescending(x=>x.pubtime).First();
-            ViewBag.ArticleId = CArticle.id;
-            var comments = (from c in db.Comments where(c.ArticleId==CArticle.id) select c).ToArray();
-            ViewBag.Comments = comments;
-
-            return View();
-        }
+        
 
        
-        [Authorize]
-        [HttpPost]
-        public ActionResult Index(Models.Comment comment)
+        // GET: /My/
+        [HttpGet]
+        [Authorize(Roles="Admin")]
+        public ActionResult Index()
         {
+
+            string[] roles = Roles.GetRolesForUser();
+           //  var a=Roles.GetAllRoles();
+            if (User.IsInRole("Admin"))
+                ViewBag.HelloAdmin = "Hello "+User.Identity.Name;
+
+            ViewBag.UserName = User.Identity.Name;
             
-            if (ModelState.IsValid)
-            {
-            
-                db.Comments.Add(comment);
-                var dbusers = new UsersContext();
-                var CurrentUser = dbusers.UserProfiles.Find(WebSecurity.GetUserId(User.Identity.Name));
-                if (CurrentUser.Points == null)
-                    CurrentUser.Points = 1;
-                else
-                    CurrentUser.Points++;
-                dbusers.Entry(CurrentUser).State = EntityState.Modified;
-                dbusers.SaveChanges();
-                db.SaveChanges();
-
-                return RedirectToAction("Index");
-
-            }
-
             return View();
         }
 
+
         [Authorize(Roles="Admin")]
         [HttpGet]
-        public ActionResult CreateComment() {
+        public ActionResult CreateComment() {//create a comment page 
             ViewBag.UserName = User.Identity.Name;
             return View();
         }
 
         [Authorize(Roles="Admin")]
         [HttpPost]
-        public ActionResult CreateComment(Models.Comment comment)
+        public ActionResult CreateComment(Models.Comment comment)//post a comment to the db
         {
             
             if (ModelState.IsValid)
@@ -79,14 +60,54 @@ namespace Shauli_blog.Controllers
 
             }
             return CreateComment();
+ }
+
+
+
+        public ActionResult SearchCommentAdmin(long ?id,long? AId,string name, string description, string website, string email, DateTime? Sdate, DateTime? Fdate, short? point)
+        {
+            //search by 
+
+            var comments =
+            (from c in db.Comments where c.ArticleId == AId select c);
+          
+            if (id != null && comments.ToArray().Length != 0)
+                comments = (from c in db.Comments where c.id==id select c).Intersect(comments);
+            else if (id != null)
+                comments = (from c in db.Comments where c.id==id select c);
+
+           if (name != "" && comments.ToArray().Length != 0)
+                comments = (from c in db.Comments where c.description.Contains(name) select c).Intersect(comments);
+            else if (name != "")
+                comments = (from c in db.Comments where c.name.Contains(name) select c);
+
+
+            if (description != "" && comments.ToArray().Length != 0)
+                comments = (from c in db.Comments where c.description.Contains(description) select c).Intersect(comments);
+            else if (description != "")
+                comments = (from c in db.Comments where c.description.Contains(description) select c);
+            if (email != "" && comments.ToArray().Length != 0)
+                comments = (from c in db.Comments where c.email == email select c).Intersect(comments);
+            else if (email != "")
+                comments = (from c in db.Comments where c.email == email select c);
+            if (website != "" && comments.ToArray().Length != 0)
+                comments = (from c in db.Comments where c.website == website select c).Intersect(comments);
+            else if (website != "")
+                comments = (from c in db.Comments where c.website == website select c);
+            if (Sdate != null && Fdate != null && comments.ToArray().Length != 0)
+                comments = (from c in db.Comments where c.pubTime >= Sdate && c.pubTime <= Fdate select c).Intersect(comments);
+            else if (Sdate != null && Fdate != null)
+                comments = (from c in db.Comments where c.pubTime >= Sdate && c.pubTime <= Fdate select c);
+
+            var finalcomments = comments.ToArray();
+
+            return View(finalcomments);
 
         }
-
         
 
-
-        
         public ActionResult SearchComment(string name,string description,string website,string email,DateTime ?Sdate,DateTime ?Fdate, short ?point){
+            //search by 
 
             var comments =
                 (from c in db.Comments where c.name == name select c);
@@ -115,7 +136,7 @@ namespace Shauli_blog.Controllers
         }
 
         [Authorize(Roles = "Admin")]
-        public ActionResult ViewAll() {
+        public ActionResult ViewAll() {//list all of the comments
 
             return View(db.Comments.ToArray());
         
@@ -123,14 +144,14 @@ namespace Shauli_blog.Controllers
         }
 
         [Authorize(Roles="Admin")]
-        public ActionResult ViewComment(long id) {
+        public ActionResult ViewComment(long id) {// view the comment by id
             
             var comment = db.Comments.Find(id);
             return View(comment);
         }
 
         [Authorize(Roles = "Admin")]
-        public ActionResult Delete(long id)
+        public ActionResult Delete(long id)//delete a Comment by id 
         {
             
             var comment = db.Comments.Find(id);
@@ -175,16 +196,98 @@ namespace Shauli_blog.Controllers
         public ActionResult ShowUsers() {
             var db = new UsersContext();
             var Users = db.UserProfiles.ToArray();
-
+           
             return View(Users);
         }
+        public ActionResult PartialList(string role)
+        {
+            var query = db.UserProfiles;
+            if (role == null ||role=="All")
+            {
+
+
+                return PartialView(query);
+            }
+            else
+            {
+                var Requested = Roles.GetUsersInRole(role);
+
+
+                var query2 = query.Where(s => Requested.Contains(s.UserName));
+                return PartialView(query2);
+            }
+
+
+
+        }
+
+
+
+
+
+
+        [Authorize(Roles = "Admin")]
+        public ActionResult EditUserRoles(long id=-1)
+        {
+            if (id != -1)
+            {
+                var User = db.UserProfiles.Find(id);
+                ViewBag.Role = Roles.GetRolesForUser(User.UserName).ToArray()[0];
+                return View(User);
+            }
+            else return RedirectToAction("Index");
+        }
+
+        [Authorize(Roles = "Admin")]
+        [HttpPost]
+        public ActionResult EditUserRoles(UserProfile User,string newRole)
+        {
+            if (ModelState.IsValid)
+            {
+                string usercurrentrole = Roles.GetRolesForUser(User.UserName).ElementAt(0);
+                Roles.RemoveUserFromRole(User.UserName,usercurrentrole);
+                if (newRole=="Admin") {
+                    Roles.AddUserToRole(User.UserName, "Admin");
+                }
+                else if (newRole=="Contractor") {
+                    Roles.AddUserToRole(User.UserName, "Contractor");
+                }
+                else if (newRole=="User") {
+                    Roles.AddUserToRole(User.UserName, "User");
+                
+                }
+                return RedirectToAction("ShowUsers");
+            }
+            else return RedirectToAction("Index");
+        }
+
         [Authorize(Roles="Admin")]
         public ActionResult DeleteUser(long id) {
-            var db = new UsersContext();
+            
             var User = db.UserProfiles.Find(id);
-            db.UserProfiles.Remove(User);
-            db.SaveChanges();
+            if (User != null)
+            {
+                var UserRole=Roles.GetRolesForUser(User.UserName).ElementAt(0);
+                Roles.RemoveUserFromRole(User.UserName, UserRole);
+            
+                db.UserProfiles.Remove(User);
+                db.SaveChanges();
+                
+            }
             return RedirectToAction("ShowUsers");
+        }
+
+        public ActionResult UserDetails(long id) {
+
+            var User = db.UserProfiles.Find(id);
+            if (User != null){
+                ViewBag.Role = Roles.GetRolesForUser(User.UserName).ElementAt(0);
+                return View(User);  
+            
+            
+            }
+            else
+                return RedirectToAction("Index");
         }
